+++ lib/language_pack/rails7.rb
@@ -8,8 +8,7 @@ class LanguagePack::Rails7 < LanguagePack::Rails6
     rails_version = bundler.gem_version('railties')
     return false unless rails_version
     is_rails = rails_version >= Gem::Version.new('7.a') &&
-      rails_version < Gem::Version.new('8.0.0')
+      rails_version < Gem::Version.new('8.a')
     return is_rails
   end
 end
-